requirements.txt
Flask~=3.0.3
python-dotenv~=1.0.1
openai~=1.50.2
requests~=2.32.3
streamlit~=1.38.0